Introducing device-plugin configuration
This PR is adding a configmap that contains the configuration of the
device plugin. The configmap map is built during the deploy command
using kustomizer. The configuration parameter are:
namespace to deploy into, plugin unix socket name, number of devices
to report on the node, prefix for devices' IDs that will be reported to
the kubelet, prefixes for annotations/env variable that will be added
to the allocated container, as well as prefix of the devices to be mapped
from the host into container